Refactor: Refactored Preact Components
